feat: renaming Token and TokenInfo objects and changing graph-client accordingly
