+++ packages/apps/graph-client/src/components/token-table/token-table.tsx
@@ -1,4 +1,4 @@
-import type { Token } from '@/__generated__/sdk';
+import type { NonFungibleTokenBalance } from '@/__generated__/sdk';
 import {
   Box,
   Cell,
@@ -13,7 +13,7 @@ import { atoms } from '@kadena/react-ui/styles';
 import React from 'react';
 
 interface ITokenTableProps {
-  tokens: Token[];
+  tokens: NonFungibleTokenBalance[];
 }
 
 export const TokenTable = (props: ITokenTableProps): JSX.Element => {
