+++ packages/apps/graph-client/src/pages/account/overview/[fungible]/[account]/[chain].tsx
@@ -2,7 +2,7 @@ import type {
   FungibleChainAccountTransactionsConnection,
   FungibleChainAccountTransfersConnection,
   NonFungibleChainAccountTransactionsConnection,
-  Token,
+  NonFungibleTokenBalance,
 } from '@/__generated__/sdk';
 import {
   useGetFungibleChainAccountQuery,
@@ -321,7 +321,7 @@ const ChainAccount: React.FC = () => {
                   <TokenTable
                     tokens={
                       nonFungibleChainAccountData.nonFungibleChainAccount
-                        .nonFungibles as Token[]
+                        .nonFungibles as NonFungibleTokenBalance[]
                     }
                   />
                 </GridItem>
