+++ packages/apps/graph/src/graph/index.ts
@@ -9,9 +9,9 @@ import './objects/guard';
 import './objects/miner-key';
 import './objects/non-fungible-account';
 import './objects/non-fungible-chain-account';
+import './objects/non-fungible-token';
+import './objects/non-fungible-token-balance';
 import './objects/signer';
-import './objects/token';
-import './objects/token-info';
 import './objects/transaction';
 import './objects/transaction-capability';
 import './objects/transaction-command';
