+++ packages/apps/graph/src/graph/objects/non-fungible-chain-account.ts
@@ -9,7 +9,7 @@ import { builder } from '../builder';
 import { tokenDetailsLoader } from '../data-loaders/token-details';
 import type { NonFungibleChainAccount } from '../types/graphql-types';
 import { NonFungibleChainAccountName } from '../types/graphql-types';
-import Token from './token';
+import NonFungibleTokenBalance from './non-fungible-token-balance';
 
 export default builder.node(
   builder.objectRef<NonFungibleChainAccount>(NonFungibleChainAccountName),
@@ -42,7 +42,7 @@ export default builder.node(
       chainId: t.exposeID('chainId'),
       accountName: t.exposeString('accountName'),
       nonFungibles: t.field({
-        type: [Token],
+        type: [NonFungibleTokenBalance],
         complexity: COMPLEXITY.FIELD.PRISMA_WITHOUT_RELATIONS,
         async resolve(parent) {
           try {
