+++ packages/apps/graph/src/graph/objects/non-fungible-token.ts
@@ -1,6 +1,6 @@
 import { builder } from '../builder';
 
-export default builder.objectType('TokenInfo', {
+export default builder.objectType('NonFungibleToken', {
   description: 'Information related to a token.',
   fields: (t) => ({
     supply: t.exposeInt('supply'),
